V0.0.02A
Separado login/senha e tela de boas vindas para otimizar uso de recursos, antiga versão apresentava lentidão.
Ordern de execução UI>Login1>Senha
1TesteCodigo: apenas para testar blocos de código sem alterar o código fonte
Botões com wx: versão anterior mas de uso futuro/base
